chore(auth): refactor chain to coin and code cleanup
