+++ src/renderer/account/components/Management/EncryptedInput/index.js
@@ -1,28 +1,22 @@
 import { compose, withState } from 'recompose';
-import { withData, withActions, progressValues } from 'spunky';
+import { withData } from 'spunky';
 
 import withConfirm from 'shared/hocs/withConfirm';
 import { withErrorToast } from 'shared/hocs/withToast';
-import { DEFAULT_CHAIN } from 'shared/values/chains';
-import withProgressChange from 'shared/hocs/withProgressChange';
+import { DEFAULT_COIN } from 'shared/values/coins';
 
 import authActions from 'auth/actions/authActions';
 
 import EncryptedInput from './EncryptedInput';
 
-const { FAILED } = progressValues;
-
 const mapAuthDataToProps = (account) => ({ account });
-const mapAddAccountActionsToProps = (actions) => ({
-  addAccount: (data) => actions.call(data)
-});
 
 // TODO Clean up
 export default compose(
   withConfirm(),
   withErrorToast(),
   withState('passphrase', 'setPassphrase', ''),
-  withState('chainType', 'setChainType', DEFAULT_CHAIN),
+  withState('chainType', 'setChainType', DEFAULT_COIN),
   withState('data', 'setData', ({ encryptedData }) => encryptedData),
   withData(authActions, mapAuthDataToProps),
 )(EncryptedInput);
